intDriver: Add timeout WA for lost IRQs
Have a 10 seconds timeout by default after which the code checks
if the PCI device has an interrupts pending in its status register.
If this happens twice, return zero to user-space as if an interrupt
has occured as a workaround for lost MSIs.
On timeout, also check if there are signals pending, and if so return
an error code indicating interruption by signal, to support graceful
shutdown.

Signed-off-by: Elad Nachman <enachman@marvell.com>